Implementado protección de Endpoints

+++ backend/controllers/tareaController.js
@@ -2,7 +2,7 @@ const asyncHandler = require('express-async-handler')
 const Tarea = require('../models/tareaModel')
 
 const getTareas = asyncHandler( async (req, res) => {
-    const tareas = await Tarea.find()
+    const tareas = await Tarea.find({user: req.user.id})
     res.status(200).json(tareas)
 })
 
@@ -15,7 +15,8 @@ const setTareas = asyncHandler( async (req, res) => {
     }
 
     const tareas = await Tarea.create({
-        texto : req.body.texto
+        texto : req.body.texto,
+        user : req.user.id
     })
 
     res.status(201).json(tareas)
@@ -26,9 +27,16 @@ const updateTareas = asyncHandler( async (req, res) => {
 
     const tarea = await Tarea.findById(req.params.id)
 
+    // Verificamos que la tarea exista
     if(!tarea){
         res.status(400)
         throw new Error('Tarea no encontrada')
+    }
+
+    // Verificamos que la tarea pertenece al usuario del token
+    if(tarea.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('ACCESO NO AUTORIZADO, La tarea no pertenece al usuario logeado')
     }
 
     const tareaModificada = await Tarea.findByIdAndUpdate(req.params.id, req.body, {new: true} )
@@ -45,7 +53,11 @@ const deleteTareas = asyncHandler( async (req, res) => {
         throw new Error('Tarea no encontrada')
     }
     
-    // await tarea.remove
+    // Verificamos que la tarea pertenece al usuario del token
+    if(tarea.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('ACCESO NO AUTORIZADO, La tarea no pertenece al usuario logeado')
+    }
 
     const tareaEliminada = await Tarea.findByIdAndDelete(req.params.id)
